Moved homeScreen.tsx back to client so that in hopes of not causing a merge conflict
