+++ main-src/src/client/homeScreen.tsx
@@ -1,5 +1,5 @@
 import "./components/homeScreen.css"
-import {NavBar} from "./NavBar";
+import {NavBar} from "./components/NavBar";
 
 const Home = () => {
     return (
